refactor: InputValidator 패키지 util로 변경 및 util에 private 생성자 추가
