+++ src/main/java/lotto/util/LottoNumberGenerator.java
@@ -5,6 +5,8 @@ import static camp.nextstep.edu.missionutils.Randoms.pickUniqueNumbersInRange;
 import java.util.List;
 
 public class LottoNumberGenerator {
+    private LottoNumberGenerator() {}
+
     public static List<Integer> generate() {
         return pickUniqueNumbersInRange(1, 45, 6);
     }
